Add budget, hours, and link to FAQ doc.
Someone I shared the post with raised really good points about the justice issues involved with not posting the pay rate here, and I really agree. (Great article they referenced here: http://nonprofitaf.com/2015/06/when-you-dont-disclose-salary-range-on-a-job-posting-a-unicorn-loses-its-wings/) We've been sharing the hours commitment, budget, and FAQ doc with people as they reach out to us or when directly sharing the post with people, so in the interest of fairness I think it should be directly on the job description for anyone who is interested.